+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next version
 * Add `::all()` to the kind type to iterate over all kind variants
+* Generate customizable implementation of `Display` trait
 
 ## v0.0.3 - 2023-08-05
 * Make generated `kind()` function public